==> notebooks/03_financial_fraud/0.0-gy-env-check.ipynb <==
Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.

SPDX-License-Identifier: Apache-2.0


Recommended env configuration following the instruction in **[TADDY](https://github.com/yuetan031/TADDY_pytorch)**

### Requirments for TADDY
* Python==3.8
* PyTorch==1.7.1
* Transformers==3.5.1
* Scipy==1.5.2
* Numpy==1.19.2
* Networkx==2.5
* Scikit-learn==0.23.2



import sys 
sys.path.append('../../src/anomaly_detection_spatial_temporal_data/')

%load_ext autoreload
%autoreload 2

from project_setting import TADDY

taddy_env = TADDY()
